Improvements after test day 1: - delay for sail angle to match delay of the wind sensor - apparent wind angle filter time constant increased from 1 to 4s. - resulting increase of stabilization time after a maneuver to 3s   in ReferenceValues. - For optimization of AOA better logging. - one minor printf type fix. - suppression of that logging output during simulate_ship tests - update of the fake_boat model to reflect what we found out about   measurement delays - longer time constant requires adaption in filter block test

+++ trunk/modem/modemd_main.cc
@@ -142,7 +142,7 @@ int main(int argc, char** argv) {
     // Check for new SMS messages.
     list<Modem::SmsMessage> messages;
     if (m.ReceiveSMSMessages(&messages) == Modem::OK) {
-      syslog(LOG_INFO, "Received SMS messages: %d messages.", messages.size());
+      syslog(LOG_INFO, "Received SMS messages: %d messages.", (int)messages.size());
     } else {
       syslog(LOG_ERR, "Receive SMS messages: command failed!");
     }
